create transition between usage statistic and gas station detail, rename packages and create new
