+++ app/src/main/AndroidManifest.xml
@@ -69,7 +69,10 @@
         <activity
             android:name="com.SmartJeagers.CheckFuel.VolumeCompareBad"
             android:screenOrientation="locked" />
-        <activity android:name="com.SmartJeagers.CheckFuel.utils.AuthenticationManager" />
+        <activity android:name="com.SmartJeagers.CheckFuel.managers.AuthenticationManager" />
+        <activity
+            android:name="com.SmartJeagers.CheckFuel.GasStationInfo"
+            android:screenOrientation="locked" />
 
         <meta-data
             android:name="preloaded_fonts"
