+++ app/src/main/java/com/SmartJeagers/CheckFuel/managers/DatabaseManaferForProblem.java
@@ -1,4 +1,4 @@
-package com.SmartJeagers.CheckFuel.utils;
+package com.SmartJeagers.CheckFuel.managers;
 
 import android.util.Log;
