+++ app/src/main/java/com/SmartJeagers/CheckFuel/managers/HistoryManager.java
@@ -1,7 +1,9 @@
-package com.SmartJeagers.CheckFuel.utils;
+package com.SmartJeagers.CheckFuel.managers;
 
 import android.content.Context;
+import android.os.Build;
 
+import androidx.annotation.RequiresApi;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
@@ -13,6 +15,7 @@ import java.util.List;
 
 public class HistoryManager {
 
+    @RequiresApi(api = Build.VERSION_CODES.N)
     public static void setConfig(RecyclerView recyclerView, Context context, List<Refill> refills, List<DayOfUse> daysOfUse) {
 
         recyclerView.setLayoutManager(new LinearLayoutManager(context));
@@ -20,4 +23,3 @@ public class HistoryManager {
     }
 
 }
-
