+++ app/src/main/java/com/SmartJeagers/CheckFuel/models/OnItemClickListener.java
@@ -1,0 +1,5 @@
+package com.SmartJeagers.CheckFuel.models;
+
+public interface OnItemClickListener {
+    void onItemClick(ItemStatistic itemStatistic);
+}
