+++ app/src/main/res/layout/item_of_list_usage_statistic.xml
@@ -81,14 +81,15 @@
         android:layout_width="28dp"
         android:layout_height="28dp"
         android:layout_marginEnd="8dp"
+        android:contentDescription="@string/todo"
+        android:onClick="goToInfo"
         android:rotation="270"
         android:visibility="visible"
         app:layout_constraintBottom_toTopOf="@+id/line"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.506"
-        app:srcCompat="@drawable/gray_orientation"
-        android:contentDescription="@string/todo" />
+        app:srcCompat="@drawable/gray_orientation" />
 
     <View
         android:id="@+id/line"
